# docs/swagger.yaml
basePath: /api/v1
definitions:
  base.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  controllers.CreateDebitCard.createDebitCardRequest:
    properties:
      border_color:
        type: string
      color:
        type: string
      issuer:
        type: string
      name:
        type: string
    required:
    - issuer
    - name
    type: object
...
  controllers.GetUserGreeting.getUserGreetingResponse:
    properties:
      message:
        type: string
    type: object
...
  controllers.UpdateDebitCard.updateDebitCardRequest:
    properties:
      border_color:
        type: string
      color:
        type: string
      name:
        type: string
    type: object
  controllers.UpdateUserGreeting.updateUserGreetingRequest:
    properties:
      message:
        type: string
    type: object
  controllers.UpdateUserGreeting.updateUserGreetingResponse:
    properties:
      message:
        type: string
    type: object
...
  models.DebitCardWithDetails:
    properties:
      border_color:
        type: string
      card_id:
        description: DebitCard fields
        type: string
      color:
        description: DebitCardDesign fields
        type: string
      created_at:
        type: string
      issuer:
        description: DebitCardDetail fields
        type: string
      name:
        type: string
      number:
        type: string
      status:
        description: DebitCardStatus fields
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  models.User:
    properties:
      created_at:
        type: string
      deleted_at:
        description: for soft delete
        type: string
      name:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    required:
    - name
    - user_id
    type: object
host: localhost:8080
info:
  contact:
    email: suthasinee.ler@mail.com
    name: API Support
  description: This is an auto-generated API Docs.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Backend Developer Assignment API
  version: "1.0"
paths:
...
  /debit-cards:
    get:
      description: List all debit cards for a user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.DebitCardWithDetails'
            type: array
      security:
      - ApiKeyAuth: []
      summary: List debit cards
      tags:
      - Debit Cards
    post:
      description: Create a new debit card with all its details
      parameters:
      - description: Card details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.CreateDebitCard.createDebitCardRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.DebitCardWithDetails'
      security:
      - ApiKeyAuth: []
      summary: Create debit card
      tags:
      - Debit Cards
  /debit-cards/{id}:
    delete:
      description: Delete a debit card
      parameters:
      - description: Card ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Card deleted
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Delete debit card
      tags:
      - Debit Cards
    get:
      description: Get a specific debit card by ID
      parameters:
      - description: Card ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DebitCardWithDetails'
      security:
      - ApiKeyAuth: []
      summary: Get debit card
      tags:
      - Debit Cards
    put:
      description: Update an existing debit card
      parameters:
      - description: Card ID
        in: path
        name: id
        required: true
        type: string
      - description: Card details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.UpdateDebitCard.updateDebitCardRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DebitCardWithDetails'
      security:
      - ApiKeyAuth: []
      summary: Update debit card
      tags:
      - Debit Cards
...
  /user/greeting:
    get:
      consumes:
      - application/json
      description: Retrieves the greeting message for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: Returns the greeting message
          schema:
            $ref: '#/definitions/controllers.GetUserGreeting.getUserGreetingResponse'
        "401":
          description: Unauthorized - Invalid or missing token
          schema:
            $ref: '#/definitions/base.ErrorResponse'
        "404":
          description: User greeting not found
          schema:
            $ref: '#/definitions/base.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get user's greeting message
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Update the greeting message of the authenticated user
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.UpdateUserGreeting.updateUserGreetingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Returns the updated greeting message
          schema:
            $ref: '#/definitions/controllers.UpdateUserGreeting.updateUserGreetingResponse'
        "400":
          description: Bad request - Invalid input format
          schema:
            $ref: '#/definitions/base.ErrorResponse'
        "401":
          description: Unauthorized - Invalid or missing token
          schema:
            $ref: '#/definitions/base.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/base.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/base.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update user's greeting message
      tags:
      - User
  /user/profile:
    get:
      consumes:
      - application/json
      description: Retrieves the information of the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: Returns the user information
          schema:
            $ref: '#/definitions/models.User'
        "401":
          description: Unauthorized - Invalid or missing token
          schema:
            $ref: '#/definitions/base.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/base.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get user's information
      tags:
      - User
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"